GraphQL: fix per-game achievements
Stats for individual games always return `null` through graphql due to
the dynamic keys (names), so an array is required.

The name of each game can be found under a `name` field for each game,
rather than as a key in an object.

It seems that this is the preferred method over having to request
achievements for each game separately.

Also fixed the type for `tiered` (should possibly have the same
object-to-array as above, but I have changed it for the untyped `JSON`)
and renamed `points` to `points_total` to be consistent with the REST
API.